# static/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APIbase — Universal API Hub for AI Agents",
    "version": "1.0.0",
    "description": "APIbase aggregates, normalizes, and provides APIs from hundreds of businesses in a unified format optimized for AI agent consumption.",
    "contact": {
      "name": "APIbase",
      "url": "https://apibase.pro"
    }
  },
  "servers": [
    {
      "url": "https://apibase.pro",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/v1/tools/{toolId}": {
      "post": {
        "operationId": "invokeTool",
        "summary": "Invoke a tool",
        "description": "Execute a tool invocation. Returns 402 Payment Required with x402 payment details if payment is needed.",
        "parameters": [
          {
            "name": "toolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tool execution result"
          },
          "402": {
            "description": "Payment required (x402)"
          }
        }
      }
    },
...
  },
...
}